Fim do projeto

+++ src/services/OnlinePaymentService.java
@@ -3,6 +3,6 @@ package services;
 public interface OnlinePaymentService {
 
 	Double paymentFee(Double amount);
-	Integer interest(Double amount, Integer months);
+	Double interest(Double amount, Integer months);
 	
 }
